Deleted unnecessary numMessages parameter for RPCSendImagePiece. Changed another ReadByte to ReadInt32 in PlayerControlPatch.

+++ DoomScroll/Patches/PlayerControlPatch.cs
@@ -194,7 +194,7 @@ namespace Doom_Scroll.Patches
                         DoomScroll._log.LogInfo("--------------\nReceiving RPC image\n--------------");
                         int numMessages = reader.ReadInt32();
                         byte pID = reader.ReadByte();
-                        int imgID = reader.ReadByte();
+                        int imgID = reader.ReadInt32();
                         DoomScrollImage currentImage = new DoomScrollImage(numMessages, pID, imgID);
                         string currentImageKey = $"{pID}{imgID}";
                         currentImagesAssembling.Add((string)$"{pID}{imgID}", currentImage);
